+++ modules/component/spec/core/cd-aware/cd-aware_creator.spec.ts
@@ -6,6 +6,7 @@ import {
   NextObserver,
   Observer,
   of,
+  throwError,
   Unsubscribable,
 } from 'rxjs';
 
@@ -22,7 +23,7 @@ import {
 class CdAwareImplementation<U> implements OnDestroy {
   public renderedValue: any = undefined;
   public error: any = undefined;
-  public completed: boolean = false;
+  public completed = false;
   private readonly subscription: Unsubscribable;
   public cdAware: CdAware<U | undefined | null>;
   resetContextObserver: NextObserver<any> = {
@@ -144,13 +145,11 @@ describe('CdAware', () => {
     });
 
     it('error handling', () => {
+      cdAwareImplementation.cdAware.nextPotentialObservable(
+        throwError('Error!')
+      );
       expect(cdAwareImplementation.renderedValue).toBe(undefined);
-      cdAwareImplementation.cdAware.subscribe({
-        error: (e: Error) => expect(e).toBeDefined(),
-      });
-      expect(cdAwareImplementation.renderedValue).toBe(undefined);
-      // @TODO use this line
-      // expect(cdAwareImplementation.error).toBe(ArgumentNotObservableError);
+      expect(cdAwareImplementation.error).toBe('Error!');
       expect(cdAwareImplementation.completed).toBe(false);
     });