+++ modules/router-store/migrations/8_0_0/index.ts
@@ -5,12 +5,12 @@ import {
   createReplaceChange,
   visitTSSourceFiles,
   commitChanges,
-} from '@ngrx/router-store/schematics-core';
+} from '../../schematics-core';
 
 function updateRouterStoreImport(): Rule {
   return (tree: Tree) => {
     visitTSSourceFiles(tree, (sourceFile) => {
-      let changes: ReplaceChange[] = [];
+      const changes: ReplaceChange[] = [];
       ts.forEachChild(sourceFile, function findDecorator(node) {
         if (!ts.isDecorator(node)) {
           ts.forEachChild(node, findDecorator);