+++ modules/schematics/src/ngrx-push-migration/index.ts
@@ -10,7 +10,7 @@ import {
   visitNgModuleExports,
   addImportToModule,
   addExportToModule,
-} from '@ngrx/schematics/schematics-core';
+} from '../../schematics-core';
 
 const ASYNC_REGEXP = /\| {0,}async/g;
 const REACTIVE_MODULE = 'ReactiveComponentModule';
@@ -27,13 +27,14 @@ export function migrateToNgrxPush(): Rule {
   return (host: Tree) =>
     visitTemplates(host, (template) => {
       let match: RegExpMatchArray | null;
-      let changes: Change[] = [];
+      const changes: Change[] = [];
       while ((match = ASYNC_REGEXP.exec(template.content)) !== null) {
         const m = match.toString();
 
         changes.push(
           new ReplaceChange(
             template.fileName,
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
             template.start + match.index!,
             m,
             m.replace('async', 'ngrxPush')