+++ modules/store/schematics/ng-add/index.ts
@@ -1,3 +1,5 @@
+import * as ts from 'typescript';
+import { Path, dirname } from '@angular-devkit/core';
 import {
   Rule,
   SchematicContext,
@@ -25,9 +27,7 @@ import {
   addPackageToPackageJson,
   platformVersion,
   parseName,
-} from '@ngrx/store/schematics-core';
-import { Path, dirname } from '@angular-devkit/core';
-import * as ts from 'typescript';
+} from '../../schematics-core';
 import { Schema as RootStoreOptions } from './schema';
 
 function addImportToNgModule(options: RootStoreOptions): Rule {
@@ -115,6 +115,66 @@ function addNgRxStoreToPackageJson() {
   };
 }
 
+function addNgRxESLintPlugin() {
+  return (host: Tree, context: SchematicContext) => {
+    const eslintConfigPath = '.eslintrc.json';
+    const docs =
+      'https://github.com/timdeschryver/eslint-plugin-ngrx/#eslint-plugin-ngrx';
+
+    const eslint = host.read(eslintConfigPath)?.toString('utf-8');
+    if (!eslint) {
+      return host;
+    }
+
+    addPackageToPackageJson(
+      host,
+      'devDependencies',
+      'eslint-plugin-ngrx',
+      '^1.0.0'
+    );
+    context.addTask(new NodePackageInstallTask());
+
+    try {
+      const json = JSON.parse(eslint);
+      if (json.overrides) {
+        json.overrides
+          .filter((override: { files?: string[] }) =>
+            override.files?.some((file: string) => file.endsWith('*.ts'))
+          )
+          .forEach(configureESLintPlugin);
+      } else {
+        configureESLintPlugin(json);
+      }
+
+      host.overwrite(eslintConfigPath, JSON.stringify(json, null, 2));
+
+      context.logger.info(`
+The NgRx ESLint Plugin is installed and configured with the recommended config.
+
+If you want to change the configuration, please see ${docs}.
+`);
+      return host;
+    } catch (err) {
+      context.logger.warn(`
+Something went wrong while adding the NgRx ESLint Plugin.
+The NgRx ESLint Plugin is installed but not configured.
+
+Please see ${docs} to configure the NgRx ESLint Plugin.
+
+Details:
+${err.message}
+`);
+    }
+
+    return host;
+  };
+}
+
+function configureESLintPlugin(json: any): void {
+  json.plugins = [...(json.plugins || []), 'ngrx'];
+  json.extends = [...(json.extends || []), 'plugin:ngrx/recommended'];
+}
+
 export default function (options: RootStoreOptions): Rule {
   return (host: Tree, context: SchematicContext) => {
     options.path = getProjectPath(host, options);
@@ -142,7 +202,7 @@ export default function (options: RootStoreOptions): Rule {
     }
 
     const templateSource = apply(url('./files'), [
-      filter((_) => (options.minimal ? false : true)),
+      filter(() => (options.minimal ? false : true)),
       applyTemplates({
         ...stringUtils,
         ...options,
@@ -156,6 +216,7 @@ export default function (options: RootStoreOptions): Rule {
         chain([addImportToNgModule(options), mergeWith(templateSource)])
       ),
       options && options.skipPackageJson ? noop() : addNgRxStoreToPackageJson(),
+      options && options.skipESLintPlugin ? noop() : addNgRxESLintPlugin(),
     ])(host, context);
   };
 }