+++ modules/store/src/reducer_manager.ts
@@ -1,6 +1,5 @@
 import { Inject, Injectable, OnDestroy, Provider } from '@angular/core';
 import { BehaviorSubject, Observable } from 'rxjs';
-
 import { ActionsSubject } from './actions_subject';
 import {
   Action,
@@ -20,11 +19,16 @@ export abstract class ReducerObservable extends Observable<
   ActionReducer<any, any>
 > {}
 export abstract class ReducerManagerDispatcher extends ActionsSubject {}
-export const UPDATE = '@ngrx/store/update-reducers' as '@ngrx/store/update-reducers';
+export const UPDATE = '@ngrx/store/update-reducers' as const;
 
 @Injectable()
-export class ReducerManager extends BehaviorSubject<ActionReducer<any, any>>
+export class ReducerManager
+  extends BehaviorSubject<ActionReducer<any, any>>
   implements OnDestroy {
+  get currentReducers(): ActionReducerMap<any, any> {
+    return this.reducers;
+  }
+
   constructor(
     private dispatcher: ReducerManagerDispatcher,
     @Inject(INITIAL_STATE) private initialState: any,